changes ml_score field to decimal type

+++ db/migrate/20230803141537_add_ml_score_to_points.rb
@@ -1,5 +1,5 @@
 class AddMlScoreToPoints < ActiveRecord::Migration[5.2]
   def change
-    add_column :points, :ml_score, :integer
+    add_column :points, :ml_score, :decimal
   end
 end
